splitting the large test file to make it easier to work with
